Problem 4 solved

+++ src/main/shared/Common.java
@@ -2,15 +2,14 @@ package main.shared;
 
 public class Common {
 
-	
-	public static boolean checkIsDivisible(int x, int d) {
-		if ( (x % d) == 0) {
+	public static boolean checkIsDivisible(long x, long d) {
+		if ((x % d) == 0) {
 			return true;
 		}
 		return false;
 	}
 
-	public static boolean isEven(int x) {
+	public static boolean isEven(long x) {
 		return checkIsDivisible(x, 2);
 	}
 }
